[os] Move leftover code from `setup.sh`
Move leftover `update content` related code into the
`update_content.sh` file.

Ref: https://github.com/alrra/dotfiles/commit/c201bbf8623f4e3b4cb64967e8a58bf5c85585f7

+++ src/os/update_content.sh
@@ -19,11 +19,15 @@ main() {
 
         print_in_purple "\n Update content\n\n"
 
-        execute \
-            "git fetch --all \
-                && git reset --hard origin/master \
-                && git clean -fd" \
-            "Update content"
+        ask_for_confirmation "Do you want to update the content from the 'dotfiles' directory?"
+
+        if answer_is_yes; then
+            execute \
+                "git fetch --all \
+                    && git reset --hard origin/master \
+                    && git clean -fd" \
+                "Update content"
+        fi
 
     fi
